% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_murderdata}
\alias{get_murderdata}
\title{Get data from \href{http://murderdata.org}{The Murder Accountability Project}}
\usage{
get_murderdata(data_type = c("clearances", "supplementary"))
}
\arguments{
\item{data_type}{Do you want the \code{clearences} or the \code{supplementary}
data set? The function defaults to the clearences data set.}
}
\value{
a tibble
}
\description{
The Murder Accountability Project is a nonprofit group organized in 2015 and
dedicated to educate Americans on the importance of accurately accounting for
unsolved homicides within the United States. They seek to obtain information
from federal, state and local governments about unsolved homicides and to
publish this information. The Project’s Board of Directors is composed of
retired law enforcement investigators, investigative journalists, criminologists
and other experts on various aspects of homicide.
}
\details{
With data from the Murder Accountability Project you can determine how often
police departments in american communities clear a homicide through arrest using the
\code{clearances} data. You can also explore individual cases reported to the
FBI or obtained by the Murder Accountability Project under local Freedom of
Information Acts. Using the \code{supplementary} data, you can look for patterns in
the occurrence of specific types of homicides and how often police identified
the offender.

Homicide investigators may find this site useful in testing theories about
murders in their community. The \code{supplementary} data available is especially
useful to test theories about suspects who may have killed across multiple
jurisdictions or within the same jurisdiction over a period of time.
The \code{supplementary} data is the most complete accounting of homicides available
anywhere.
}
\examples{
clearances <- get_murderdata()
dplyr::glimpse(clearances)

supplementary <- get_murderdata("supplementary")
dplyr::glimpse(supplementary)

}
