% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_by_country}
\alias{sk_by_country}
\title{Serial Killers by Country}
\format{A tibble with 20 rows and 5 variables:
\describe{
  \item{country}{Country name}
  \item{no_sk}{Number of Serial Killers}
  \item{pct_sk}{Percent of Serial Killers}
  \item{share_world_pop}{Country's Share of World Population}
  \item{ratio}{Ratio}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_by_country
}
\description{
Caution should be taken in reviewing this table as the number of serial killers by country is a function
of the number of actual killers, the number caught, and the number reported by police or the media
}
\details{
Note: Serial killers operating in multiple countries are not included

Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
