% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_by_race_decade}
\alias{sk_by_race_decade}
\title{U.S. Serial Killers Percentage by Race and Decade}
\format{A tibble with 13 rows and 6 variables:
\describe{
  \item{decade}{Decade}
  \item{white}{White race}
  \item{black}{Black race}
  \item{hispanic}{Hispanic race}
  \item{asian}{Asian race}
  \item{native_american}{Native American race}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_by_race_decade
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
