% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_definition_of}
\alias{sk_definition_of}
\title{Number of Separate Serial Killers Operating in a Given Year in the U.S.}
\format{A tibble with 95 rows and 3 variables:
\describe{
  \item{year}{year}
  \item{two_or_more}{Definition of having 2 or more victims}
  \item{three_or_more}{Definition of having 3 or more victims}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_definition_of
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
