% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_gender_decade_all}
\alias{sk_gender_decade_all}
\title{Frequency of Serial Killers Gender by Decade: U.S. and International}
\format{A tibble with 24 rows and 4 variables:
\describe{
  \item{decade}{Decade}
  \item{gender}{Gender of serial killers}
  \item{n}{Total count}
  \item{percent}{Percentage}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_gender_decade_all
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
