% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_iq_overview}
\alias{sk_iq_overview}
\title{Serial Killer IQ - overview}
\format{A tibble with 5 rows and 3 variables:
\describe{
  \item{area}{Overall area pertaining to the serial killer}
  \item{variable}{Variable connected to area}
  \item{iq}{IQ}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_iq_overview
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
