% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_method}
\alias{sk_method}
\title{Method by Which Victim Was Killed}
\format{A tibble with 12 rows and 3 variables:
\describe{
  \item{method}{Method of Death}
  \item{n}{Total count}
  \item{pct}{Percentage}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_method
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
