% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_motives}
\alias{sk_motives}
\title{Broad Motives}
\format{A tibble with 10 rows and 7 variables:
\describe{
  \item{motive}{Motive of the serial killer}
  \item{n}{Total count}
  \item{percentage}{Percentage}
  \item{white}{White race, US only}
  \item{black}{Black race, US only}
  \item{hisp}{Hispanic race, US only}
  \item{asian}{Asian race, US only}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_motives
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
