% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_victims_demographics}
\alias{sk_victims_demographics}
\title{Characteristics of U.S. Serial Killer Victims}
\format{A tibble with 116 rows and 4 variables:
\describe{
  \item{variable}{Overarching demographic variable}
  \item{value}{Value of overarching demographic variable}
  \item{n}{Total count}
  \item{pct}{Percentage}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_victims_demographics
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
