% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_victims_state}
\alias{sk_victims_state}
\title{Victims By State}
\format{A tibble with 51 rows and 5 variables:
\describe{
  \item{state}{US state}
  \item{n}{Total count}
  \item{victim_rank}{Rank by victim count}
  \item{pop_rank}{Rank by population}
  \item{diffence}{Difference between victim rank and population rank}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_victims_state
}
\description{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
