% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_victims_yearly}
\alias{sk_victims_yearly}
\title{Number of Serial Killer Victims by Year}
\format{A tibble with 116 rows and 3 variables:
\describe{
  \item{year}{Year of the murder}
  \item{us}{Count for the United States}
  \item{canada}{Count for Canada}
}}
\source{
Radford University/FGCU Serial Killer Database
}
\usage{
sk_victims_yearly
}
\description{
Note: The numbers for 2014 and 2015 should be interpreted with caution as
there is normally a lag between a murder and the identification of the serial killer.
}
\details{
Aamodt, M. G. (2015, November 23). Serial killer statistics. Retrieved (2017-12-12)
from http://maamodt.asp.radford.edu/serial killer information center/project description.htm
}
\keyword{datasets}
